First commit baby!
The README contains what I want this project to have. Of course
this might change, however I do think that these are the minimum
requirements for an artist's webpage. I looked at Seong-Jin Cho's
page and thought: "That's nice, I could do that", so here we go.
The plan is to finish this in under two months, but I'm aiming for
the first prototype to be done before the 15/12/24.