Create page header prototype
Understanding the background-image property was a little tricky. It took
me awhile to realize what box-sizing: border-box is for. I reorganized
the project a little, specfically the static files.

I learned that some of the things I did manually, like scaling images,
is done automatically by Wordpress. Not that I care. Haha.

I need to do a number of things, such as prepare another image, decide
whether the blockquote placement can be improved, and make the hamburger
bigger on larger devices (and its neighboring elements).