Finish brief-biography prototype
Things went faster now that I understand how background images
work. However I am having difficulty getting the wrapper to account
for the border & padding of the <a> element.

Anyways, I decided not to get fixated on this one issue and carry
on with the next, very cruicial part of the site:

                Upcoming performances!!