Include page title and minor contact fixes
I decided to place the page title close to the top right corner for
both the mobile-tablet and desktop viewports. It clashed too much
with the nav section because of the text styles (e.g. kern).

I also included the "border" fix for the contact section, I just
noticed it had the same issue that the performances section had
previously.

I think this is it for the home page prototype. I'll later make
changes to the upcoming performances section  but this is enough
for now. I don't think there's much to do for the inner pages. I'll
do my best to work smart and fast.