Docker init
I made use of cargo-chef to speed up future builds times. However
I had issues with the clux/muslrust image that worked for amd64, but
did not seem to work for arm64. For now, I am sticking to building for
amd64 as that is what my VPS is currently running on.

I will later explore the clux/muslrust documentation for arm64 builds.