Fix Dockerfile and .yaml
First and foremost, I removed the deploy section inside the .yaml,
because I had not yet configured Docker swarm. I believe it was causing
issues with 'actually' running the service on port 8080.

Furthermore, I learned the askama counts the templates directory and its
contents as dependencies, as the rendered templates are baked into the
binary. With that, I added only the /static directory as volume via the
.yaml file.

Lastly, I revised a line in the .dockerignore file to ignore the entire
/static directory.