Initial setup
To start off the Go rewrite, I decided to opt for a cmd/ internal/
project structure. So far I have found it simple to organize shared
files that way.

I also learned about centralizing application-level components for
explicit dependency injection with struct fields and constructors.

This feels nice, I also heard it should help with testing later on,
I believe this.

Now I've only started with the base.html and home.html pages, but
it's a start. I also still need to complete the startup logic for
the App constructor.

I'll do my best to keep pushing forward. It's been a rough past
few days- but fudge it, we ball.