Fix usage of http/template
Embed the templates directory, then populate the map using a
separate helper function.

I parsed the base.html file, and assign it to `base`. I then clone
this `base` variable to build each template.Template instance.

Separate `base` instances are needed, becase we populate the blocks
with different things, based on which page it is for.