Implement session management
Many thanks to Mohamed Said at themsaid.com, I definitely would not
have figured this out on my own.

https://themsaid.com/building-secure-session-manager-in-go

This implementation of session management is separated into three
parts that have their own roles:

1. the Session struct

This type has an id and a data field. It also has two timestamp
fields that enable us to expire sessions based on inactivity or an
arbitrary duration. The data field is essentially a catch-all field
for user specific data.

The Session type has methods for accessing/manipulating session
data. Right now there are only three basic methods, but we might
add more later.

In this implementation, sessions are request-scoped variables. That
means that only goroutine has access to the session at a time. This
is accomplished by making sure our handlers do not spawn additional
goroutines, so that we can pass our session through the context.

2. the SessionStore interface (and in-memory implementation)

The interface defines methods for accessing/manipulating sessions
themselves. It's recommended in production settings to use some
kind of persistent storage, but for now I've stuck to an in-memory,
map solution.

The interface also defines a garbage collection method.
More on this later.

3. the SessionManager struct

The manager is instantiated on startup. With a constructor function
we pass our SessionStore and other configuration values.

The SessionManager has it's own garbage collection method, which
uses a ticker to periodically call the SessionStore's garbage
collection method, at a designated interval. This method is invoked
in it's constructor function.

The type really only exposes one public method, i.e. Handle()
It's a function that satisfies the http.Handler interface, which
makes it essentially middleware.

The session middleware prepares the session, adds it to the context
then wraps the ResponseWriter to track writes and ensure the cookie
is properly included in the response.

---

I've got a lot to do still, for example authentication. However I
am really happy that I found this article as a starting point.

Keep it up!